requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.

pbr>=1.6 # Apache-2.0

eventlet
flask
sqlalchemy
keystoneauth
oslo.cache
oslo.config
oslo.messaging
oslo.log
oslo.db
python-keystoneclient
requests
six
uwsgi
